requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.44.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
djongo==1.3.3
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.2
